requirements.txt
python-telegram-bot[job-queue]==20.3
groq-sdk>=0.22.0
python-dotenv==0.21.0
sqlalchemy==1.4.39
pytz==2022.1
psutil>=5.8.0